requirements.txt
streamlit
youtube-transcript-api
google-genai
python-dotenv
fpdf2
beautifulsoup4
requests
Pillow
groq